fix: use dunce::canonicalize for PathGuard and path equality
Windows std::fs::canonicalize returns \\?\ UNC paths that break
Path::starts_with against non-UNC roots and look wrong in agent
messages. bline hit this as a patchloom embedder.

Add containment::safe_canonicalize (dunce), use it for PathGuard
roots, ancestor walks, extra roots, plan cwd checks, symlink
resolve, and rename same-file equality. Extract containment tests
to containment_tests.rs. Regression tests lock no UNC prefix and
AllowIfContained absolute paths.

Closes #1931
