+++ src/containment.rs
@@ -53,6 +53,21 @@
 
 use std::path::{Component, Path, PathBuf};
 
+/// Canonicalize a path without the Windows `\\?\` UNC prefix when safe.
+///
+/// On Windows, [`std::fs::canonicalize`] returns paths like
+/// `\\?\C:\Users\...`. Those break [`Path::starts_with`] when compared to a
+/// non-UNC root (or vice versa) and look wrong in agent-facing display.
+/// [`dunce::canonicalize`] strips the prefix when the path is shorter than
+/// `MAX_PATH` and has no special characters. On non-Windows this is a
+/// transparent passthrough to `std::fs::canonicalize`.
+///
+/// Prefer this for PathGuard roots, containment checks, and any path that may
+/// be compared with `starts_with` or shown in error messages.
+pub fn safe_canonicalize(path: &Path) -> std::io::Result<PathBuf> {
+    dunce::canonicalize(path)
+}
+
 /// Policy for handling absolute paths.
 #[derive(Debug, Clone, PartialEq, Eq)]
 pub enum AbsolutePathPolicy {
@@ -216,7 +231,7 @@ impl PathGuard {
                 AbsolutePathPolicy::AllowAdditionalRoots(extra) => {
                     let mut allowed = vec![self.canon_root.clone()];
                     for r in extra {
-                        if let Ok(c) = r.canonicalize() {
+                        if let Ok(c) = safe_canonicalize(r) {
                             allowed.push(c);
                         }
                     }
@@ -308,12 +323,10 @@ impl PathGuard {
         root: PathBuf,
         policy: AbsolutePathPolicy,
     ) -> Result<Self, ContainmentError> {
-        let canon_root = root
-            .canonicalize()
-            .map_err(|e| ContainmentError::Canonicalize {
-                path: root.display().to_string(),
-                source: e,
-            })?;
+        let canon_root = safe_canonicalize(&root).map_err(|e| ContainmentError::Canonicalize {
+            path: root.display().to_string(),
+            source: e,
+        })?;
         Ok(Self {
             root,
             canon_root,
@@ -429,6 +442,9 @@ fn normalize_lexical(path: &Path) -> PathBuf {
 
 /// Canonicalize a path, or if it doesn't exist, canonicalize the nearest
 /// existing ancestor and append the remaining components.
+///
+/// Uses [`safe_canonicalize`] (dunce) so Windows UNC prefixes do not break
+/// containment `starts_with` checks against the PathGuard root.
 fn canonicalize_or_ancestor(path: &Path) -> std::io::Result<PathBuf> {
     // Normalize `..` and `.` lexically first so the ancestor walk never
     // encounters components that `file_name()` would silently skip.
@@ -436,7 +452,7 @@ fn canonicalize_or_ancestor(path: &Path) -> std::io::Result<PathBuf> {
     let path = normalized.as_path();
 
     if path.exists() {
-        return path.canonicalize();
+        return safe_canonicalize(path);
     }
     // Walk up to the nearest existing ancestor.
     let mut ancestor = path;
@@ -448,7 +464,7 @@ fn canonicalize_or_ancestor(path: &Path) -> std::io::Result<PathBuf> {
                 if let Some(file_name) = ancestor.file_name() {
                     tail_components.push(file_name.to_os_string());
                 }
-                let canon_ancestor = p.canonicalize()?;
+                let canon_ancestor = safe_canonicalize(p)?;
                 // Rebuild the path by appending tail components in reverse.
                 let mut result = canon_ancestor;
                 for c in tail_components.into_iter().rev() {
@@ -462,7 +478,7 @@ fn canonicalize_or_ancestor(path: &Path) -> std::io::Result<PathBuf> {
                 }
                 ancestor = p;
             }
-            None => return path.canonicalize(),
+            None => return safe_canonicalize(path),
         }
     }
 }
@@ -476,489 +492,5 @@ const _: () = {
 };
 
 #[cfg(test)]
-mod tests {
-    use super::*;
-    use std::fs;
-
-    #[test]
-    fn relative_path_within_workspace() {
-        let dir = tempfile::TempDir::new().unwrap();
-        fs::write(dir.path().join("file.txt"), "ok").unwrap();
-        let guard = PathGuard::new(dir.path().to_path_buf(), AbsolutePathPolicy::Reject).unwrap();
-        guard.check_path("file.txt").unwrap();
-    }
-
-    #[test]
-    fn relative_path_with_safe_parent_traversal() {
-        let dir = tempfile::TempDir::new().unwrap();
-        fs::create_dir_all(dir.path().join("src")).unwrap();
-        fs::write(dir.path().join("Cargo.toml"), "ok").unwrap();
-        let guard = PathGuard::new(dir.path().to_path_buf(), AbsolutePathPolicy::Reject).unwrap();
-        // src/../Cargo.toml stays within workspace
-        guard.check_path("src/../Cargo.toml").unwrap();
-    }
-
-    #[test]
-    fn relative_path_escaping_workspace() {
-        let dir = tempfile::TempDir::new().unwrap();
-        let guard = PathGuard::new(dir.path().to_path_buf(), AbsolutePathPolicy::Reject).unwrap();
-        let err = guard.check_path("../../etc/passwd").unwrap_err();
-        assert!(matches!(err, ContainmentError::Escaped { .. }));
-    }
-
-    #[test]
-    fn absolute_path_with_allow_if_contained() {
-        let dir = tempfile::TempDir::new().unwrap();
-        fs::write(dir.path().join("inside.txt"), "ok").unwrap();
-        let guard = PathGuard::new(
-            dir.path().to_path_buf(),
-            AbsolutePathPolicy::AllowIfContained,
-        )
-        .unwrap();
-        let abs = dir.path().join("inside.txt");
-        guard.check_path(abs.to_str().unwrap()).unwrap();
-    }
-
-    /// Return an absolute path string that is guaranteed outside any temp dir.
-    fn outside_absolute_path() -> &'static str {
-        #[cfg(unix)]
-        {
-            "/etc/passwd"
-        }
-        #[cfg(windows)]
-        {
-            "C:\\Windows\\System32\\notepad.exe"
-        }
-    }
-
-    #[test]
-    fn absolute_path_outside_workspace_with_allow_if_contained() {
-        let dir = tempfile::TempDir::new().unwrap();
-        let guard = PathGuard::new(
-            dir.path().to_path_buf(),
-            AbsolutePathPolicy::AllowIfContained,
-        )
-        .unwrap();
-        let err = guard.check_path(outside_absolute_path()).unwrap_err();
-        assert!(matches!(err, ContainmentError::Escaped { .. }));
-    }
-
-    #[test]
-    fn absolute_path_with_reject_policy() {
-        let dir = tempfile::TempDir::new().unwrap();
-        let guard = PathGuard::new(dir.path().to_path_buf(), AbsolutePathPolicy::Reject).unwrap();
-        let err = guard.check_path(outside_absolute_path()).unwrap_err();
-        assert!(matches!(err, ContainmentError::AbsolutePath(_)));
-    }
-
-    #[cfg(unix)]
-    #[test]
-    fn symlink_escaping_workspace() {
-        let dir = tempfile::TempDir::new().unwrap();
-        let link = dir.path().join("escape");
-        std::os::unix::fs::symlink("/tmp", &link).unwrap();
-        let guard = PathGuard::new(dir.path().to_path_buf(), AbsolutePathPolicy::Reject).unwrap();
-        let err = guard.check_path("escape").unwrap_err();
-        assert!(matches!(err, ContainmentError::Escaped { .. }));
-    }
-
-    #[test]
-    fn nonexistent_file_in_existing_directory() {
-        let dir = tempfile::TempDir::new().unwrap();
-        let guard = PathGuard::new(dir.path().to_path_buf(), AbsolutePathPolicy::Reject).unwrap();
-        let result = guard.check_path("new_file.txt");
-        assert!(
-            result.is_ok(),
-            "non-existent file with safe ancestor should be allowed"
-        );
-    }
-
-    #[test]
-    fn empty_path() {
-        let dir = tempfile::TempDir::new().unwrap();
-        let guard = PathGuard::new(dir.path().to_path_buf(), AbsolutePathPolicy::Reject).unwrap();
-        // Empty string resolves to cwd itself, which is contained.
-        guard.check_path("").unwrap();
-    }
-
-    #[test]
-    fn single_parent_rejected() {
-        let dir = tempfile::TempDir::new().unwrap();
-        let guard = PathGuard::new(dir.path().to_path_buf(), AbsolutePathPolicy::Reject).unwrap();
-        guard.check_path("..").expect_err("expected error");
-    }
-
-    #[test]
-    fn deep_traversal_rejected() {
-        let dir = tempfile::TempDir::new().unwrap();
-        let guard = PathGuard::new(dir.path().to_path_buf(), AbsolutePathPolicy::Reject).unwrap();
-        guard
-            .check_path("a/b/../../../escape")
-            .expect_err("expected error");
-    }
-
-    #[test]
-    fn dot_relative_path_allowed() {
-        let dir = tempfile::TempDir::new().unwrap();
-        fs::create_dir_all(dir.path().join("foo")).unwrap();
-        fs::write(dir.path().join("foo/bar.json"), "{}").unwrap();
-        let guard = PathGuard::new(dir.path().to_path_buf(), AbsolutePathPolicy::Reject).unwrap();
-        guard.check_path("./foo/bar.json").unwrap();
-    }
-
-    #[cfg(unix)]
-    #[test]
-    fn new_file_through_symlink_dir_rejected() {
-        let dir = tempfile::TempDir::new().unwrap();
-        let link = dir.path().join("link_dir");
-        std::os::unix::fs::symlink("/tmp", &link).unwrap();
-        let guard = PathGuard::new(dir.path().to_path_buf(), AbsolutePathPolicy::Reject).unwrap();
-        let err = guard.check_path("link_dir/new_file.txt").unwrap_err();
-        assert!(
-            matches!(err, ContainmentError::Escaped { .. }),
-            "new file through symlink escaping workspace should be rejected"
-        );
-    }
-
-    #[test]
-    fn check_path_returns_canonicalized_path() {
-        let dir = tempfile::TempDir::new().unwrap();
-        fs::write(dir.path().join("test.txt"), "ok").unwrap();
-        let guard = PathGuard::new(dir.path().to_path_buf(), AbsolutePathPolicy::Reject).unwrap();
-        let resolved = guard.check_path("test.txt").unwrap();
-        // The returned path should be absolute (canonicalized).
-        assert!(resolved.is_absolute());
-        assert!(resolved.ends_with("test.txt"));
-    }
-
-    #[test]
-    fn root_and_canon_root_accessors() {
-        let dir = tempfile::TempDir::new().unwrap();
-        let guard = PathGuard::new(dir.path().to_path_buf(), AbsolutePathPolicy::Reject).unwrap();
-        assert_eq!(guard.root(), dir.path());
-        assert!(guard.canon_root().is_absolute());
-    }
-
-    #[test]
-    fn new_with_nonexistent_root_returns_canonicalize_error() {
-        let err = PathGuard::new(
-            PathBuf::from("/nonexistent_patchloom_test_dir_xyz"),
-            AbsolutePathPolicy::Reject,
-        )
-        .unwrap_err();
-        assert!(matches!(err, ContainmentError::Canonicalize { .. }));
-        let msg = err.to_string();
-        assert!(
-            msg.contains("failed to canonicalize"),
-            "expected canonicalize message, got: {msg}"
-        );
-    }
-
-    #[test]
-    fn error_display_absolute_path() {
-        let err = ContainmentError::AbsolutePath("/etc/passwd".to_string());
-        let msg = err.to_string();
-        assert!(
-            msg.contains("absolute paths are not allowed") && msg.contains("/etc/passwd"),
-            "unexpected message: {msg}"
-        );
-    }
-
-    #[test]
-    fn error_display_escaped() {
-        let err = ContainmentError::Escaped {
-            path: "../../secret".to_string(),
-            root: "/home/user".to_string(),
-        };
-        let msg = err.to_string();
-        assert!(
-            msg.contains("escapes workspace")
-                && msg.contains("../../secret")
-                && msg.contains("/home/user"),
-            "unexpected message: {msg}"
-        );
-    }
-
-    #[test]
-    fn parent_escape_error_includes_workspace_root() {
-        let dir = tempfile::TempDir::new().unwrap();
-        let guard = PathGuard::new(dir.path().to_path_buf(), AbsolutePathPolicy::Reject).unwrap();
-        let err = guard.check_path("../escape.txt").unwrap_err();
-        let msg = err.to_string();
-        let root_disp = dir.path().display().to_string();
-        assert!(
-            msg.contains("escapes") && msg.contains(&root_disp),
-            "escape error must name workspace root {root_disp:?}, got: {msg}"
-        );
-        assert!(
-            !msg.contains("(workspace: )"),
-            "must not print empty workspace root: {msg}"
-        );
-    }
-
-    #[test]
-    fn error_source_canonicalize_returns_inner_error() {
-        use std::error::Error;
-        let io_err = std::io::Error::new(std::io::ErrorKind::NotFound, "gone");
-        let err = ContainmentError::Canonicalize {
-            path: "test".to_string(),
-            source: io_err,
-        };
-        // strengthened (was bare is_some); expect gives context on failure
-        let _src = err
-            .source()
-            .expect("Canonicalize error must carry inner io source");
-        let msg = err.to_string();
-        assert!(msg.contains("failed to canonicalize"), "got: {msg}");
-    }
-
-    #[test]
-    fn error_source_non_canonicalize_returns_none() {
-        use std::error::Error;
-        let err = ContainmentError::AbsolutePath("/x".to_string());
-        assert!(err.source().is_none());
-        let err2 = ContainmentError::Escaped {
-            path: "..".to_string(),
-            root: "/r".to_string(),
-        };
-        assert!(err2.source().is_none());
-    }
-
-    #[test]
-    fn allow_additional_roots_allows_temp_and_extra() {
-        let dir = tempfile::TempDir::new().unwrap();
-        let temp = std::env::temp_dir();
-        let guard = PathGuard::new(
-            dir.path().to_path_buf(),
-            AbsolutePathPolicy::AllowAdditionalRoots(vec![temp.clone()]),
-        )
-        .unwrap();
-        // absolute in extra root should work
-        let tmp_file = temp.join("patchloom_test_extra.txt");
-        // create it
-        std::fs::write(&tmp_file, "ok").unwrap();
-        let res = guard.check_path(tmp_file.to_str().unwrap());
-        res.unwrap();
-        // clean
-        let _ = std::fs::remove_file(&tmp_file);
-    }
-
-    #[test]
-    fn allow_additional_roots_still_rejects_outside() {
-        let dir = tempfile::TempDir::new().unwrap();
-        let guard = PathGuard::new(
-            dir.path().to_path_buf(),
-            AbsolutePathPolicy::allow_workspace_and_temp_dir(),
-        )
-        .unwrap();
-        let err = guard.check_path("/etc/passwd").unwrap_err();
-        assert!(matches!(err, ContainmentError::Escaped { .. }));
-    }
-
-    #[test]
-    fn builder_allows_temp() {
-        let dir = tempfile::TempDir::new().unwrap();
-        let guard = PathGuard::builder(dir.path().to_path_buf())
-            .allow_temp_directory()
-            .build()
-            .unwrap();
-        // should allow temp
-        let temp = std::env::temp_dir().join("patchloom_builder_test.txt");
-        std::fs::write(&temp, "ok").unwrap();
-        guard.check_path(temp.to_str().unwrap()).unwrap();
-        let _ = std::fs::remove_file(&temp);
-    }
-
-    /// Regression test for #781: literal `/tmp/...` (common spelling from agents/LLMs)
-    /// must be allowed by `allow_temp_directory()` even on macOS where
-    /// std::env::temp_dir() is under /var/folders and /tmp symlinks to /private/tmp.
-    #[cfg(unix)]
-    #[test]
-    fn builder_allows_literal_tmp_path_unix() {
-        let dir = tempfile::TempDir::new().unwrap();
-        let guard = PathGuard::builder(dir.path().to_path_buf())
-            .allow_temp_directory()
-            .build()
-            .unwrap();
-
-        // Use a unique name under the conventional /tmp to simulate real usage
-        // (host experiment-mode paths, agent-generated paths, etc.).
-        let tmp_path = format!("/tmp/patchloom_781_test_{}.txt", std::process::id());
-        let _ = std::fs::remove_file(&tmp_path);
-        std::fs::write(&tmp_path, "data for 781").unwrap();
-        let res = guard.check_path(&tmp_path);
-        assert!(
-            res.is_ok(),
-            "literal /tmp path must be allowed: {:?}",
-            res.err()
-        );
-        let _ = std::fs::remove_file(&tmp_path);
-    }
-
-    /// Non-existing file under /tmp must still be allowed (via ancestor canonicalization).
-    #[cfg(unix)]
-    #[test]
-    fn builder_allows_nonexistent_under_literal_tmp() {
-        let dir = tempfile::TempDir::new().unwrap();
-        let guard = PathGuard::builder(dir.path().to_path_buf())
-            .allow_temp_directory()
-            .build()
-            .unwrap();
-
-        let tmp_path = format!("/tmp/patchloom_781_nonexist_{}.txt", std::process::id());
-        // do not create it
-        let res = guard.check_path(&tmp_path);
-        assert!(
-            res.is_ok(),
-            "non-existing /tmp path must be allowed via ancestor: {:?}",
-            res.err()
-        );
-    }
-
-    #[test]
-    fn allow_workspace_and_temp_dir_allows_std_temp() {
-        let dir = tempfile::TempDir::new().unwrap();
-        let guard = PathGuard::new(
-            dir.path().to_path_buf(),
-            AbsolutePathPolicy::allow_workspace_and_temp_dir(),
-        )
-        .unwrap();
-        let temp = std::env::temp_dir().join("patchloom_awtd_std.txt");
-        std::fs::write(&temp, "ok").unwrap();
-        guard.check_path(temp.to_str().unwrap()).unwrap();
-        let _ = std::fs::remove_file(&temp);
-    }
-
-    /// #781: allow_workspace_and_temp_dir must also cover literal /tmp.
-    #[cfg(unix)]
-    #[test]
-    fn allow_workspace_and_temp_dir_allows_literal_tmp() {
-        let dir = tempfile::TempDir::new().unwrap();
-        let guard = PathGuard::new(
-            dir.path().to_path_buf(),
-            AbsolutePathPolicy::allow_workspace_and_temp_dir(),
-        )
-        .unwrap();
-
-        let tmp_path = format!("/tmp/patchloom_781_awtd_{}.txt", std::process::id());
-        let _ = std::fs::remove_file(&tmp_path);
-        std::fs::write(&tmp_path, "data").unwrap();
-        guard.check_path(&tmp_path).unwrap();
-        let _ = std::fs::remove_file(&tmp_path);
-    }
-
-    #[test]
-    fn would_allow_works() {
-        let dir = tempfile::TempDir::new().unwrap();
-        let guard = PathGuard::new(dir.path().to_path_buf(), AbsolutePathPolicy::Reject).unwrap();
-        assert!(guard.would_allow("foo.txt"));
-        assert!(!guard.would_allow("../escape"));
-    }
-
-    /// Basic coverage for the helper (addresses review note for direct test of system_temp...).
-    #[test]
-    fn system_temp_directory_roots_contains_expected_entries() {
-        let roots = system_temp_directory_roots();
-        // Always at least the std one
-        assert!(!roots.is_empty());
-        // On unix we explicitly add the conventional ones
-        #[cfg(unix)]
-        {
-            assert!(roots.iter().any(
-                |r| r == std::path::Path::new("/tmp") || r == std::path::Path::new("/var/tmp")
-            ));
-            // std env temp is present
-            let stdt = std::env::temp_dir();
-            assert!(roots.iter().any(|r| r == &stdt));
-        }
-    }
-
-    /// Regression for coverage: paths using ../ to escape an *allowed additional root*
-    /// (e.g. literal /tmp on macOS) must still be rejected after canonicalization.
-    /// This was not explicitly asserted in prior temp-allow tests.
-    #[cfg(unix)]
-    #[test]
-    fn allow_temp_rejects_escape_from_tmp_via_parent() {
-        let dir = tempfile::TempDir::new().unwrap();
-        let guard = PathGuard::builder(dir.path().to_path_buf())
-            .allow_temp_directory()
-            .build()
-            .unwrap();
-
-        let escape_path = "/tmp/../etc/passwd";
-        let err = guard.check_path(escape_path).unwrap_err();
-        assert!(
-            matches!(err, ContainmentError::Escaped { .. }),
-            "expected Escaped for escape from /tmp root via .., got {:?}",
-            err
-        );
-    }
-
-    /// Regression: `..` components in non-existent absolute paths must not be
-    /// silently dropped during ancestor canonicalization. `Path::file_name()`
-    /// returns `None` for `..`, so the old code skipped those components,
-    /// making a path like `workspace/nonexistent/../../outside/secret.txt`
-    /// appear to be `workspace/nonexistent/outside/secret.txt` (inside).
-    #[test]
-    fn dotdot_in_nonexistent_absolute_path_rejected() {
-        let base = tempfile::TempDir::new().unwrap();
-        let workspace = base.path().join("workspace");
-        let outside = base.path().join("outside");
-        fs::create_dir(&workspace).unwrap();
-        fs::create_dir(&outside).unwrap();
-        fs::write(outside.join("secret.txt"), "sensitive").unwrap();
-
-        let guard =
-            PathGuard::new(workspace.clone(), AbsolutePathPolicy::AllowIfContained).unwrap();
-
-        // workspace/nonexistent/../../outside/secret.txt
-        // resolves lexically to: base/outside/secret.txt (outside workspace)
-        let escape = workspace
-            .join("nonexistent")
-            .join("..")
-            .join("..")
-            .join("outside")
-            .join("secret.txt");
-
-        let result = guard.check_path(escape.to_str().unwrap());
-        assert!(
-            matches!(result, Err(ContainmentError::Escaped { .. })),
-            "path with .. through non-existent dir should be rejected, got: {:?}",
-            result,
-        );
-    }
-
-    /// Same bypass with AllowAdditionalRoots policy.
-    #[test]
-    fn dotdot_in_nonexistent_path_rejected_with_additional_roots() {
-        let base = tempfile::TempDir::new().unwrap();
-        let workspace = base.path().join("workspace");
-        let extra_root = base.path().join("extra");
-        let outside = base.path().join("outside");
-        fs::create_dir(&workspace).unwrap();
-        fs::create_dir(&extra_root).unwrap();
-        fs::create_dir(&outside).unwrap();
-        fs::write(outside.join("data.txt"), "sensitive").unwrap();
-
-        let guard = PathGuard::new(
-            workspace.clone(),
-            AbsolutePathPolicy::AllowAdditionalRoots(vec![extra_root]),
-        )
-        .unwrap();
-
-        // extra uses nonexistent to escape via ..
-        let escape = workspace
-            .join("nonexistent")
-            .join("..")
-            .join("..")
-            .join("outside")
-            .join("data.txt");
-
-        let result = guard.check_path(escape.to_str().unwrap());
-        assert!(
-            matches!(result, Err(ContainmentError::Escaped { .. })),
-            "dotdot escape via additional-roots should be rejected, got: {:?}",
-            result,
-        );
-    }
-}
+#[path = "containment_tests.rs"]
+mod tests;
